requirements.txt
fastapi[standard]
python-dotenv
uvicorn
sqlmodel
alembic
pydantic-settings
phonenumbers
requests
pyotp